IMP - Refactor input components
