+++ src/components/elements/input/autoComplete/main.module.scss
@@ -1,4 +1,4 @@
-@use '../../../../public/colors';
+@use '../../../../../public/colors';
 
 div {
   @mixin section {
