+++ src/components/pages/gameData/skillAtk/in/sections/skill.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import {useI18n} from '../../../../../../i18n/hook';
-import {InputPanel} from '../../../../../elements/input/main';
+import {InputPanel} from '../../../../../elements/input/panel/main';
 import {SectionProps} from '../types';
 import {overrideInputData} from '../utils/inputData';
